+++ Springboot Java/src/main/java/com/api/carrental/Service/BookingService.java
@@ -28,7 +28,7 @@ public class BookingService {
 
 	public Booking createBooking(Booking booking) {
 	    int carId = booking.getCar().getId();       // uses carId from Car entity
-	    Long customerId = booking.getCustomer().getId(); // uses id from Customer entity
+	    int customerId = booking.getCustomer().getId(); // uses id from Customer entity
 	    
 	    /*fetch the car entity from database using carid ensure the car exist if not throws exception*/	    
 	    Car car = carrentalRepository.findById(carId)
